hack: add deploy script
Add a script to build the image and deploy it to the local openshift
registry (or any other registry).

By default uses the forklift-konveyor namespace and
forklift-controller:latest image name. The upload will be done to the
default openshift registry and the script will edit the CSV to use the
image from registry instead of the image from quay.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>