Add new states to wait for snapshot readiness.
Split up snapshot creation states into separate states for snapshot
creation and for waiting for the created snapshot to be ready to
transfer. This avoids starting an importer pod on an oVirt snapshot that
is guaranteed to fail with "Cannot transfer Virtual Disk. Snapshot is
currently being created for VM".

Signed-off-by: Matthew Arnold <marnold@redhat.com>