Recheck after provider periodically
Currently once a provider has been reconciled it will no longer be
checked, unless changes have been made to it. This may lead to a
situation where a provider that is no longer active will appear as
ready.

This patch reconciles every 10 minutes to make sure it is still
functional.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>