Set the CPU policy
This patch will configure the VM with the desired CPU policy. Kubevirt
supports only `dedicated`. The other policies won't be respected and
a warning should be shown for them. As for `dedicated` the nodes in the
cluster need to support `CPU manager`, otherwise the VM won't be able to
schedule.