Bug 2055201: Avoid starting importer before snapshots are ready to transfer. (#443)
* Add new states to wait for snapshot readiness.

Split up snapshot creation states into separate states for snapshot
creation and for waiting for the created snapshot to be ready to
transfer. This avoids starting an importer pod on an oVirt snapshot that
is guaranteed to fail with "Cannot transfer Virtual Disk. Snapshot is
currently being created for VM".

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Poll job status of snapshot creation command.

This is a nicer way to wait for successful snapshot creation than
polling the status of the snapshot itself. Derive a predictable value
from existing plan information so that no data structures need changing,
and hash it with MD5 so that it fits in the correlation ID length limit.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Missed handling one error case previously.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Simplify snapshot wait states.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Try to check all snapshot jobs for readiness.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

Signed-off-by: Matthew Arnold <marnold@redhat.com>