Prevent data volume deletion
With the new CDI change, the data volumes being deleted when the import
completes. However, the controller lookup for the data volumes in order
to know when the disk transfer is complete and and progress with the
migration flow. In order to prevent it, an annotation is added to the
data volumes.